I aligned the light theme with the dark
I don't actually use the light theme - at all.  So I wouldn't be
surprised in the least to learn that it is somewhat wonky.  My goal in
this commit was to take the dark theme, copy it over the light theme,
and then reverse the background colors.  In effect I refreshed the
entries.

This probably has some unwanted artifacts that I don't know about.