Definitions for YAML content
The YAML parser seems to miss a few things.  For example, if you use xed
to view a YAML file, you'll see that it properly sees escape sequences
and such. So not all of the definitions work properly.

Nonetheless, here are YAML definitions.

As an aside, I noticed that the jpg files had execute permissions.  Not
sure when that slipped in, but I set better permissions while I was
here.