Substantial do over
I realized I was doing it wrong all along when it came time to spin up a
new machine.  I copied my configuration to the new machine, and the
result looked dreadful.

There is still much work to do, but this version is much better.  To
enable this theme, add the following to your SciTEUser.properties (or
whichever config file is appropriate for your use):

SOLAR.DARK=1
import .config/scite/themes/Solarized/theme

- Note that I've since moved the theme to a folder called "themes".
I've been experimenting with other color schemes (Nord is kind of nice),
and I want to be able to switch and swap at a moment's notice.  I'm not
there yet, but using a "themes" folder helps.